refactoring and removing old endpoints
